webview: Abbreviate link text for references
We don't have any description of what each reference is specifically
about, so currently we use the URL as the link text.  This can be
quite hard to read.

Generally the most significant parts of the URL are at the beginning
(hostname) and end (query part or last path part).  Add a "urlabbrev"
filter that drops the scheme and replaces any other insignificant
part with an ellipsis.  Use it in the issue template.