Mark CVE-2019-9500 as not fixed in linux-3.16.y
CVE-2019-9500 didn't affect Linux 3.16, but the stable branch has an
incorrect backport of the fix for it that has now been reverted.
Since import_stable.py doesn't (yet) recognise and take account of
reverts, explicitly mark this as never fixed in linux-3.16.y so that
the next run won't re-add the incorrect backport here.