Generalise branch support
We currently identify branches only by name and assume that
everything except "mainline" is a stable branch and follows
the naming scheme used for stable.

Before we can support tracking of CIP's branches, we need
to remove this assumption.

Replace branch names in most places with dictionaries that
specify a short_name (the name used in issue files and in UI),
git_remote, git_name, and base_ver (the stable version that
the branch is based on).  For mainline there is no base_ver
and we will still need to treat it specially in various
places.