scripts/import_ubuntu: Accept new Mitigation field
The Ubuntu CVE tracker now records mitigation information in a
specific field instead of in general comments.  The importer
currently fails if it finds an unknown field, so we need to add
this to the list of known fields.

Possibly our own schema should be extended similarly, but I'm
not changing that right now.

Signed-off-by: Ben Hutchings <ben.hutchings@codethink.co.uk>