Use git to check if a path is really a git repo
The existing method of checking whether ".git" exists or not does not
support bare repositories, nor does it validate the path is an actual
git repository, or just an unfortunate coincidence.

Since the main purpose is to support bare repositories, use git's
method of checking if it is a bare repository or not. The command
returns "true" or "false" on stdout. If it is not a git repo, the
command fails.

Signed-off-by: Chen-Yu Tsai (Moxa) <wens@csie.org>