webview: Delete broken table container for "Open issues"
The style rule for "table-container" has a typo: as it doesn't start
with ".", it applies to a (non-existent) element type and not to a
class.

If we implement the table container properly then its content becomes
scrollable, except that the headings only move along one axis.  But the
"Open issues" page contains almost nothing but the table, so it seems
pointless and confusing to add extra scroll bars within the page.

Instead, delete the broken style rule and unwrap the table container.

Signed-off-by: Ben Hutchings <ben.hutchings@codethink.co.uk>