Prepare to use hierarchical names for comments from distributions
Currently we include comments from distribution trackers with
'Debian-' or 'Ubuntu-' as a prefix to the commenter's handle there.
The distribution name is meant to be like a namespace for those
comments, but '-' is an unusual hierarchy separator.  Use '/' as the
separator.  (I also considered using ':', but it is a special
character in YAML.)

Key names are case-sensitive, and we use lower-case for most keys, so
also use lower-case names for the distributor names.

Add a conversion script which can be run to update the existing
issues.