Load patch queue configuration along with branch configuration
Load a new configuration file, patch-queues.yml, which describes
patch queues for the kernel.  This will be used to track fixes
in stable-queue.git and distribution repositories that use patch
queues.

Currently the only required configuration key for each patch queue is
'patch_url_format', a format string used for linking to patches in the
queue.