Allow change strings to refer to a tagged version
The Debian kernel security tracker records some issues as fixed on
Debian branches without giving a specific commit or patch reference.
This indicates that they were fixed upstream and the fix was imported
into Debian.

The Ubuntu kernel security tracker records issued as fixed and
released in a particular version, but it's not always possible to
identify exactly which commit was responsible.

* Allow recording these fixes as 'version:' plus the tag name
* Add support for linking to tags in webview