prepare_remotes: Create missing remotes if the repository already exists
Currently prepare_remotes is useful for creating a repository
with all the expected remotes at one point in time.  However
if the default configuration is changed to add more remotes,
it would be useful to have a way to add those to an existing
repository.

In case the repository exists, list the existing remotes and
skip creation of any configured remotes that already exist.