prepare_remotes: Enable fetching the wanted branches and no others
For some remotes, we won't want to fetch all branches.  Check which
live branches reference the remote, and only enable fetching those.
For the 'stable' remote we continue fetching all branches.

When a remote already exists, check that all live branches are
configured to be fetched, and add any that are missing.