Introduce hierarchical names for CIP and stable branches
In preparation for adding distribution branches, rename the CIP and
stable branches to begin with 'cip/' and 'stable/' respectively.
Remove the redundant 'linux' and '.y' from them, to keep them short.

Ignore the cached stable branch information if it doesn't use the new
name format.

Add a conversion script which can be run to update the existing
issues.