import_ubuntu: Get fixed versions and look for matching git commits
In cases where issues are fixed in Ubuntu before they are fixed
upstream, the commit messages in Ubuntu's git branches won't have a
upstream commit reference, so import_stable won't be able to find
them.  However, the commit messages will often have the CVE ID.

For each issue, look up the fixed versions for each Ubuntu branch
we're tracking, then grep the git log for commits in the 90 days
before that version that mention the CVE ID.  Record the commits if we
find some; otherwise the version tag if it exists.