Add test scripts for issues found by bobfuzzer
bobfuzzer generates filesystem images and (in some cases) user-space
code to trigger filesystem crashes.  The reproducers are publicly
available on GitHub.

* Add bobfuzzer.sh, defining common functions for downloading and
  running the reproducers
* Add test script for each issue, using those functions

I excluded CVE-2019-19927 because it's actually a graphics driver
issue and I don't think it will be reliably reproducible.

I excluded CVE-2019-19037 because its filesystem image is not zipped
so would need special treatment, and the issue is apparently fixed
everywhere already.