import_debian: Fix incorrect handling of ignored issues in Debian tracker
When the import script was modified to accept fixes spanning multiple
releases, the same logic was applied to handling "ignored" statements
as well.

Turns out that doesn't work out so well, since ignored statements
only occur once, and the typing was all wrong.

Revert the handling of the ignored statement, and add more comments
explaining the logic.

Signed-off-by: Chen-Yu Tsai (Moxa) <wens@csie.org>