Add CVe-2021-22600
A double free bug in packet_set_ring() in net/packet/af_packet.c can be
exploited by a local user through crafted syscalls to escalate
privileges or deny service.
This issue was introduced by commit 61fad68 ("net/packet: tpacket_rcv:
avoid a producer race condition"). This commit was merged in 5.6.
However, it was backported to 5.4, 4.19, and 4.14 so that these kernels
are also affected but 4.4 and 4.9 are not backported.

Signed-off-by: Masami Ichikawa <masami.ichikawa@cybertrust.co.jp>