Add CVE-2020-36516
An issue was discovered in the Linux kernel through 5.16.11. The mixed
IPID assignment method with the hash-based IPID assignment policy allows
an off-path attacker to inject data into a victim's TCP session or
terminate that session.

According to the commit 23f5740 ("ipv4: avoid using shared IP generator
for connected sockets ") this bug was introduced by commit 73f156a
    ("inetpeer: get rid of ip_id_count") which was merged at 3.16-rc1.

The 4.4 kernel was fixed in its maintenance phase.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>