Add CVE-2022-0742
A remote attacker can crash victim host via malicious ICMP6 packet.
This vulnerability was introduced at commit f185de2 ("mld: add new
workqueues for process mld events").
This commit was merged in 5.13-rc1, so before this versions are not
affected by this issue.
The mainline and stable kernels are fixed.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>