Add CVE-2022-1204
An UAF bug was found in net/ax25/af_ax25.c. An attacker is able to crash system by this vulnerability.

Patches were merged in following order.

2022-03-29:	ax25: fix UAF bug in ax25_send_control(5352a761308397a0e6250fdc629bb3f615b94747) fixes 9fd75b6
2022-03-21:	ax25: Fix refcount leaks caused by ax25_cb_del(9fd75b66b8f68498454d685dc4ba13192ae069b0) fixes d01ffb9, 87563a0, and feef318
2022-02-09:	ax25: fix UAF bugs of net_device caused by rebinding operation(feef318c855a361a1eccd880f33e88c460eb63b4)
2022-02-03:	ax25: fix reference count leaks of ax25_dev(87563a043cef044fed5db7967a75741cc16ad2b1) fixes d01ffb9
2022-01-28:	ax25: add refcount in ax25_dev to avoid UAF bugs(d01ffb9eee4af165d83b08dd73ebdf9fe94a519b)

It looks route cause is commit d01ffb ("ax25: add refcount in ax25_dev to avoid UAF bugs"). It was merged in 5.17-rc3. This commit isn't backported to stable kernels. So, stable kernels seem not be affected by this CVE.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>