Add CVE-2022-1205
A Null pointer dereference and use after free in net/ax25/ax25_timer.c.
An attacker is able to crash system by this vulnerability.

This bug was introduced by commit 7ec02f5 ("ax25: fix NPD bug in
ax25_disconnect") which was merged in 5.17-rc4. This commit isn't
backported to stable kernels so that this bug only affects to mainline
kernel.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>