Add CVE-2022-0171
Bug was found in KVM SEV API that non-root user crash the host kernel by
creating confidential guest vm instance in AMD CPU that supports AMD's
SEV.

The red hat bugzilla references upstream merge commit
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=bb4ce2c65881a2b9bdcd384f54a260a12a89dd91
but it doesn't indicate which commit fixes this issue.

It seems fixed in the mainline but not sure which commit is relevant.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>