Add CVE-2022-1734
An issue was discovered in the Linux Kernel from 4.18 to 4.19, an
improper update of sock reference in TCP pacing can lead to memory/netns
leak, which can be used by remote clients.

This issue was introduced by commit 73a6bab5 ("tcp: switch pacing timer
to softirq based hrtimer") which was merged in 4.18-rc1. This patch
wasn't backported to 4.9, 4.4. So 4.4 and 4.9 aren't affected by this
issue.

This issue was fixed by commit 0a70f118 ("tcp: fix possible socket leaks
in internal pacing mode"). It said that "In linux-4.20, TCP stack
adopted EDT (Earliest Departure Time) model and this issue was
incidentally fixed.".

The import_stable script set 0a70f118475e037732557796accd0878a00fc25a as
fixed commit in the mainline.
However, it is not mentioned which commit(s) fixed this issue in the
mainline.
so, temporary sets v4.20's commit hash to fixed revision instead of
0a70f11 because commit 0a70f11 isn't in the mainline.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>