Add CVE-2022-2586
A use-after-free vulnerability was found in nf_tables. This
vulnerability may allow an attacker to escalate privilege escalation.
However, to exploit this vulnerability, it requires CAP_NET_ADMIN in
user or netns.

This bug was introduced by commit 958bee14d071 ("netfilter: nf_tables:
use new transaction infrastructure to handle sets") which was merged in
3.16-rc1. So, all stable kernels are affected by this vulnerability.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>