Add CVE-2022-2978
A freeing uninitialized memory bug was found in nilfs_mdt_destroy().
This bug occurs in an error path. If allocating memory for inode is
failed in inode_init_always(), it returns ENOMEM. Then
nilfs_mdt_destroy() is called which frees uninialized data.

It looks 4.4 kernel is affected this bug too.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>