Add CVE-2022-3630
This issues was introduced by commit 85e4ea1 ("fscache: Fix
invalidation/lookup race") in 5.19-rc6. This commit is not backported to
stable kernels so that they are not affected by this issue.
The commit 85e4ea1 is fixes d24af13 ("fscache: Implement cookie
invalidation") in 5.17-rc1. The commit d24af13 is not backported to
stable kernels too.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>