Add CVE-2023-4010
This CVE is not fixed in the mainline yet.
A reporter describes this bug on the
github(https://github.com/wanrenmi/a-usb-kernel-bug) that said
vulnerabitiy is in the usb_giveback_urb(). But that function is not
found in any kernel versions. There is a usb_giveback_urb_bh() in
drivers/usb/core/hcd.c instead.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>