Add CVE-2023-4394
This bug was introduced by commit faa775c ("btrfs: add a
btrfs_get_dev_args_from_path helper") in 5.16-rc1 and fixed in 6.0-rc3.
The commit faa775c was backported to 5.15 so that it was affected by
this vulnerability. before Linux 5.15 doesn't contain the commit so they
are not affectd.

All stable kernels and cip kernels are fixed.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>