Add CVE-2023-42754
A null pointer dereference bug was found in the ipv4_send_dest_unreach"
function in the net/ipv4/route.c.
This bug was introduced by commit ed0de45 ("ipv4: recompile ip options
in ipv4_link_failure") in 5.1-rc6.
The commit ed0de45 was backported to 4.4, 4.14, and 4.19 so these
kernels are affected as well.

To abuse this bug, an attacker requires CAP_NET_ADMIN privilege.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>