Add CVE-2022-3534
This CVE is old one. If an user checks unfixed CVEs by other
tool(e.g. cve-check in yocto project), the tool will report
CVE-2022-3534 is not fixed. So, we need to track this CVE for
the case.

Debian assigns this bug to libbpf. So cip-kernel-sec didn't
create this yaml file automatically.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>