Add CVE-2021-47487
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in  v5.15.
It seems as if this bug was introduced by commit 918698d ("drm/amd/display: Return the number of bytes parsed than allocated") in 5.10-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>