Add CVE-2021-47490
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in  v5.15.
The commit 8eb8833e ("drm/ttm: add gem base object") in 5.4-rc1 added base object into struct ttm_buffer_object.
Therefore prior to 5.4 kernels do not have an object to free in the fix.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>