Add CVE-2021-47551
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v5.16-rc2.
It seems as if this bug was introduced by commit d95e8e ("drm/amdgpu: refine create and release logic of hive info") in 5.10-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>