Add CVE-2024-36023
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc1.
It looks as if this bug was introduced by commit ac2c6375 ("orangefs: Fix sb refcount leak when allocate sb info failed.") in 5.16-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>