Add CVE-2024-36024
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc1.
It looks as if this bug was introduced by commit 8892780 ("drm/amd/display: Wake DMCUB before sending a command") in 6.8-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>