Add CVE-2024-36026
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc4.
It looks as if this bug was introduced by commit 526e6ca ("drm/amdgpu/pm: remove the repeated EnableGfxImu message sending") in 6.0-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>