Add CVE-2024-36909
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc4.
It looks as if this bug was introduced by commit ae6935e ("vmbus: split ring buffer allocation from open") in 4.20-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>