Add CVE-2024-36897
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc7.
It seems as if this bug was introduced by commit fcffbcf ("drm/amd/display: Add DCN3.1 BIOS parser support") in 5.14-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>