Add CVE-2024-36915
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc4.
It seems as if this bug was intrdocued by commit a7b75c5 ("net: pass a sockptr_t into ->setsockopt") in 5.9-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>