Add CVE-2024-36920
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc2.
It seems as if this bug was intrdocued by commit 429846b4 ("scsi: mpi3mr: Add support for MPT commands") in 5.19-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>