Add CVE-2024-36921
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.9-rc2.
It seems as if this bug was introduced by commit 87f7e2435 ("wifi: iwlwifi: mvm: add sta handling flows for MLD mode") in 6.4-rc1.
Bug introduced commit is not backported to older stable kernels.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>