scripts/report_affected: add feature to show ignored/fixed status
This commit added following features
- Show ignored and fixed status
- Output yaml format dump and save to file

To keep old behavior, above features are enabled when yaml format is
specified to --output-format option.

Following CVEs were used in testing.

CVE-2023-52479: ignored because ksmbd is not present in 4.19
CVE-2024-36977: 4.19 is not affected
CVE-2024-36971: fixed in 4.19 but not fixed in cip/4.19
CVE-2023-1390: fixed in 4.19 and cip/4.19

CVE-2023-52479 is ignored to its listed in "ignored" key.
$ ./scripts/report_affected.py --output-format=yaml  --git-repo ../kernel/ --remote-name cip:cip v4.19.312-cip109
"cip/4.19:v4.19.312-cip109":
  "affected":
    - "CVE-2024-36971"
  "fixed":
    - "CVE-2023-1390"
    - "CVE-2024-36977"
  "ignored":
    - "CVE-2023-52479"

Added "--include-ignored" option so CVE-2023-52479 is listed in affected instead of ignored.
$ ./scripts/report_affected.py --output-format=yaml --include-ignored  --git-repo ../kernel/ --remote-name cip:cip v4.19.312-cip109
"cip/4.19:v4.19.312-cip109":
  "affected":
    - "CVE-2023-52479"
    - "CVE-2024-36971"
  "fixed":
    - "CVE-2023-1390"
    - "CVE-2024-36977"
  "ignored": []

CVE-2024-36971 is ignored so that it isn't reported.
$ ./scripts/report_affected.py --git-repo ../kernel/ --remote-name cip:cip v4.19.312-cip109
cip/4.19:v4.19.312-cip109: CVE-2024-36971

Added "--include-ignored" option so CVE-2023-52479 is reported.
$ ./scripts/report_affected.py --include-ignored --git-repo ../kernel/ --remote-name cip:cip v4.19.312-cip109
cip/4.19:v4.19.312-cip109: CVE-2023-52479 CVE-2024-36971

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>