Add CVE-2024-41079
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.10-rc4.
The fabrics-cmd-auth.c was added by commit db1312dd95 ("nvmet: implement basic In-Band Authentication")
in 6.0-rc1. The fabrics-cmd.c was added by commit a07b4970f4 ("nvmet: add a generic NVMe target") in
4.8-rc1.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>