Add CVE-2024-43902
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.11-rc1.
It looks as if this bug is releated to following commits.
cdaae8371 ("drm/amd/display: Handle GPU reset for DC block") in 5.8-rc1.
dcd5fb82f ("drm/amd/display: Fix reference counting for struct dc_sink.") in 5.1-rc1
d6149086b ("Revert \"drm/amd/display: edp do not add non-edid timings\"") in.6.5-rc1.
So, 4.x are not affected.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>