Add CVE-2024-50277
There is no mention of the commit that introduced the bug in the CVE announcement.Fixed in v6.12-rc7.
It looks as if this bug was introduced by commit 74fa8f9c ("block: pass a
queue_limits argument to blk_alloc_disk") in 6.9-rc1. Since this commit,
the blk_alloc_disk returns ERR_PTR instead of NULL.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>